% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{set_proxy}
\alias{set_proxy}
\alias{unset_proxy}
\title{Set/Unset proxy settings}
\usage{
set_proxy(proxy = "127.0.0.1:7890")

unset_proxy()
}
\arguments{
\item{proxy}{\[character(1)\] The proxy link.}
}
\description{
Set/Unset proxy settings
}
